requirements.txt
google-cloud-automl==2.5.2